GameManager.kt: move to another folder
