+++ app/src/androidTest/java/ch/sdp/vibester/helper/GameManagerTest.kt
@@ -1,0 +1,4 @@
+package ch.sdp.vibester.helper
+
+class GameManagerTest {
+}
